# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JOB API",
        "description": "API da plataforma JOB para intermediação de serviços.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/trabalhos/contratacoes": {
            "get": {
                "tags": [
                    "Trabalhos"
                ],
                "summary": "Lista trabalhos contratados pelo usuário",
                "description": "Retorna os trabalhos em que o usuário autenticado é o contratante.",
                "operationId": "05cc069494050b329fd827e5cdcf0a10",
                "responses": {
                    "200": {
                        "description": "Lista de trabalhos contratados"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/trabalhos/meus": {
            "get": {
                "tags": [
                    "Trabalhos"
                ],
                "summary": "Lista trabalhos em que o usuário foi contratado",
                "description": "Retorna os trabalhos em que o usuário autenticado é o contratado.",
                "operationId": "f03135babf16aa2779d943b9d12efe58",
                "responses": {
                    "200": {
                        "description": "Lista de trabalhos do contratado"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/trabalhos/{trabalho}/interacoes": {
            "post": {
                "tags": [
                    "Trabalhos"
                ],
                "summary": "Interage com um trabalho",
                "description": "Permite ao contratado concluir ou desistir e ao contratante confirmar, contestar ou desistir conforme o status atual do trabalho.",
                "operationId": "ea911a03a93e800e791b92621c1689b2",
                "parameters": [
                    {
                        "name": "trabalho",
                        "in": "path",
                        "description": "ID do trabalho",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_interacao_tipo"
                                ],
                                "properties": {
                                    "id_interacao_tipo": {
                                        "description": "ID do tipo de interação. O tipo precisa pertencer ao contexto TRABALHO.",
                                        "type": "integer",
                                        "example": 6,
                                        "enum": [
                                            6,
                                            7,
                                            8,
                                            9,
                                            10
                                        ]
                                    },
                                    "mensagem": {
                                        "type": "string",
                                        "example": "Serviço concluído conforme combinado.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interação registrada"
                    },
                    "403": {
                        "description": "Usuário não autorizado"
                    },
                    "422": {
                        "description": "Interação não permitida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação do Laravel Sanctum",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}